Prevent crash with $info on an offline/invalid character, Added missing netdb.h include on Linux, Trade system
git-svn-id: svn://svn.eoserv.net/eoserv/trunk@97 da1c5d5e-72c9-4536-8af9-af1c9134b0e6

+++ src/util.cpp
@@ -297,7 +297,7 @@ int rand(int min, int max)
 		init = true;
 		std::srand(std::time(0));
 	}
-	return double(std::rand()) / RAND_MAX * (max - min + 1) + min;
+	return static_cast<int>(double(std::rand()) / RAND_MAX * (max - min + 1) + min);
 }
 
 }
